scripts/rag/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
pinecone>=7.0.0
openai>=1.0.0
python-dotenv>=1.0.0
